@@ -1,6 +1,6 @@
 {
   "name": "@onwidget/astrowind",
-  "version": "1.0.0-beta.43",
+  "version": "1.0.0-beta.46",
   "description": "AstroWind: A free template using Astro 4.0 and Tailwind CSS. Astro starter theme.",
   "type": "module",
   "private": true,
@@ -25,12 +25,12 @@
     "@astrojs/rss": "^4.0.7",
     "@astrojs/sitemap": "^3.1.6",
     "@astrolib/analytics": "^0.5.0",
-    "@astrolib/seo": "^1.0.0-beta.5",
+    "@astrolib/seo": "^1.0.0-beta.6",
     "@contentful/rich-text-html-renderer": "^16.6.1",
-    "@fontsource-variable/inter": "^5.0.20",
+    "@fontsource-variable/inter": "^5.1.0",
     "@iconify-json/line-md": "^1.2.1",
     "airtable": "^0.12.2",
-    "astro": "^4.15.4",
+    "astro": "^4.15.5",
     "astro-embed": "^0.7.2",
     "astro-icon": "^1.1.1",
     "contentful": "^10.12.7",
@@ -43,23 +43,23 @@
   },
   "devDependencies": {
     "@astrojs/check": "^0.9.3",
-    "@astrojs/mdx": "^3.1.5",
+    "@astrojs/mdx": "^3.1.6",
     "@astrojs/partytown": "^2.1.2",
     "@astrojs/tailwind": "5.1.0",
-    "@eslint/js": "^9.9.1",
-    "@iconify-json/flat-color-icons": "^1.1.11",
-    "@iconify-json/tabler": "^1.1.121",
+    "@eslint/js": "^9.10.0",
+    "@iconify-json/flat-color-icons": "^1.2.0",
+    "@iconify-json/tabler": "^1.2.3",
     "@tailwindcss/typography": "^0.5.15",
     "@types/eslint__js": "^8.42.3",
     "@types/js-yaml": "^4.0.9",
     "@types/lodash.merge": "^4.6.9",
     "@types/mdx": "^2.0.13",
-    "@typescript-eslint/eslint-plugin": "^8.3.0",
-    "@typescript-eslint/parser": "^8.3.0",
+    "@typescript-eslint/eslint-plugin": "^8.6.0",
+    "@typescript-eslint/parser": "^8.6.0",
     "astro-compress": "2.3.1",
-    "astro-eslint-parser": "^1.0.2",
-    "eslint": "^9.9.1",
-    "eslint-plugin-astro": "^1.2.3",
+    "astro-eslint-parser": "^1.0.3",
+    "eslint": "^9.10.0",
+    "eslint-plugin-astro": "^1.2.4",
     "globals": "^15.9.0",
     "js-yaml": "^4.1.0",
     "mdast-util-to-string": "^4.0.0",
@@ -68,9 +68,9 @@
     "reading-time": "^1.5.0",
     "sharp": "0.33.5",
     "tailwind-merge": "^2.5.2",
-    "tailwindcss": "^3.4.10",
-    "typescript": "^5.5.4",
-    "typescript-eslint": "^8.3.0",
+    "tailwindcss": "^3.4.11",
+    "typescript": "^5.6.2",
+    "typescript-eslint": "^8.6.0",
     "unist-util-visit": "^5.0.0"
   }
 }
